Rename main content components
