+++ src/components/FontAwesome/FontAwesome.component.jsx
