+++ src/pages/kontakt.js
@@ -5,7 +5,7 @@ import Navbar from "../components/Header/Navbar.component"
 import NavbarContent from "../components/Header/NavbarContent.component"
 
 // Main content
-import MainContent from "../components/Kontakt/MainContent.component"
+import MainContent from "../components/Kontakt/KontaktContent.component"
 
 // Footer
 import Footer from "../components/Footer/Footer.component"
